Tagarela_Main/requirements.txt
opencv-python
mediapipe
numpy
requests
flask
pyautogui
google-generativeai
python-dotenv
pyttsx3
SpeechRecognition
pyaudio